Wave 2 of PR changes
Added the following methods to PlaylistManager:
- ClearManager: This clears the loaded playlists as well as the child managers (if clearChildren is true). This is supposed to mimmick the functionality of refreshing in PlaylistLoaderLite, and allows the refresh button in PlaylistManager to be functional. A good use case would be if a playlist is modified externally and the cache needs to be cleared so the playlist can be updated without restarting the game.
- DeleteChildManager: Allows you to delete subfolders.
- RenameManager: Allows you to rename the folder of the current manager.

Changes to Utilities: When creating an image for a playlist without an image, only the first letter of each word is used. This is to make the images look more aesthetically pleasing while also giving enough info of what the playlist is. Additionally, the folder image is now just transparent: since PlaylistManager will now have folders (to show folders/child managers) this is a better way to distinguish between them and playlists.

LegacyPlaylistSong: Added desrialize for difficulty

All Playlists and PlaylistSongs: Made CustomData protected and now the TryGet and Set methods are needed to get/set values

This is it for the changes I need merged, I will be opening the PR soon. Please approve it in BeatMods when PlaylistManager is approved as PlaylistManager 0.3.0 will only be compatible with this lib and the older versions won't work. Likewise, old lib won't work with PlaylistManager, it has to be simultaneous.